+++ CHANGELOG.md
@@ -17,6 +17,13 @@ The format is a modified version of [Keep a Changelog](https://keepachangelog.co
 
 ## [v0.19.7.6] - 2026-06-12
 ### Improved
+- Don't start playing until external tracks are loaded and ready ([@Secozzi](https://github.com/Secozzi)) ([#160](https://github.com/quickdesh/Animiru/pull/160))
+
+### Fixed
+- Fix tracks not being selected when switching episodes ([@Secozzi](https://github.com/Secozzi)) ([#160](https://github.com/quickdesh/Animiru/pull/160))
+
+## [v0.19.7.5] - 2026-06-08
+### Improved
 - Add order priority for whitelist in track select ([@Secozzi](https://github.com/Secozzi)) ([#159](https://github.com/quickdesh/Animiru/pull/159))
 
 ### Fixed